Extract Java <-> napari conversion to new module
This way, we can keep _function.py about the Widget itself
